Don't overwrite the day script if it already exists

+++ get-input.sh
@@ -55,4 +55,4 @@ curl "https://adventofcode.com/${YEAR}/day/${DAY}/input"  \
     --cookie "$COOKIE_FILE"  \
     --compressed
 
-cp "${THIS_PATH}/blank.py" "${PADDAY}a"
+[[ ! -f "${PADDAY}a" ]] && cp "${THIS_PATH}/blank.py" "${PADDAY}a"
